==> 1_data_import.Rmd <==
---
title: "data_import"
author: "SL"
date: "5/17/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(haven)
library(Hmisc)
library(janitor)
library(skimr)
library(zoo)
library(naniar)


```

## Import data files

HSI Data: sas7bdat format
DMIS IS Data: Excel format
```{r}

setwd("C:/Users/slewa/OneDrive - cumc.columbia.edu")


persons <-
  haven::read_sas("./daily_hsi/data/R200131_persons.sas7bdat") %>%
  haven::zap_formats() %>% 
  haven::zap_label() %>%
  dplyr::mutate_at(vars(-BMI), haven::zap_empty) %>%  # convert empty character strings into missing values
  janitor::clean_names() %>% 
  mutate(height = as.numeric(height),
         weight = as.numeric(weight),
         bmi_int = bmi,
         bmi = (weight / height^2) *703)

persons

encounters <-
  haven::read_sas("./daily_hsi/data/R200131_encounters.sas7bdat") %>%
  haven::zap_formats() %>% 
  haven::zap_label() %>% # remove SAS labels
  dplyr::mutate_at(vars(-D_EVENT), haven::zap_empty) %>%  # convert empty character strings into missing values
  janitor::clean_names() %>% 
  unique()    # remove duplicate rows 


encounters

setwd("C:/Users/slewa/OneDrive - cumc.columbia.edu/risk_factors")
getwd()
```


### Join `encounters` and `persons`

```{r}

df <-
  encounters %>% 
    left_join(persons, by = "studyid") %>% 
    mutate_at(vars(studyid, dmisid, dx1, dx2, dx1_name, source, age, sex, race_ethnic, 
                   service, component, grade, hor_country, hor_state), 
            list(factor))

```



```{r}

# inspect missing values
df %>% 
  naniar::vis_miss(warn_large_data = FALSE)  


# Dataset Summary: General

Hmisc::describe(df)


```


## Initial Plots 
```{r}

# Plot all encounters by year and type

df %>%
  mutate(year = lubridate::year(d_event)) %>%
  group_by(year, source) %>%
  summarise(n = dplyr::n()) %>%
  ggplot(aes(x = year, y = n, color = source)) +
    geom_point(size = 1.5) +
    geom_line(size = 1.5) +
  theme_bw()

```

## Code for Case Definition


For individuals who meet the case definition:
• The incidence date is considered the date of the first hospitalization, outpatient medical encounter, or reportable medical event that includes a defining diagnosis of heat illness.
• An individual is considered an incident case only once per calendar year.
• For individuals with more than one heat illness diagnosis in a calendar year, heat stroke is prioritized over heat exhaustion.
• Medical encounters for each individual within a calendar year are prioritized: hospitalizations > reportable medical events > ambulatory visits.


```{r}
# Inspect multiple encounters per studyid

df %>% 
  group_by(studyid) %>% 
  count() %>% 
  arrange(desc(n)) 
df %>% 
  group_by(studyid) %>% 
  count() %>% 
  arrange(desc(n)) %>% 
  ggplot() +
    geom_bar(aes(x = n)) +
    xlim(0, 15) +
    theme_bw()

```


## Classify by HSI category (stroke, exhaustion, other)

- For Inpatient and Outpatient, use DX1/DX2 codes
- For RME, need to use "dx1_name" - alphanumeric DX codes not used after 2013

```{r}

# Classify each encounter as heat stroke, heat exhaustion, or other

# Classify RMEs

  # Inspect RME classification names

df %>% 
  dplyr::filter(source %in% "RME") %>% 
  count(dx1_name)
  
# 1 HEAT DEHYDRATION               7098  : heat_exhaustion (associated with DX 9923)
# 2 Heat Exhaustion                2442  : heat_exhaustion
# 3 HEAT EXHAUSTION                 356  : heat_exhaustion
# 4 HEAT ILLNESS                    128  : heat_exhaustion
# 5 HEAT ILLNESS, HEAT EXHAUSTION  3184  : heat_exhaustion
# 6 HEAT ILLNESS, HEAT INJURY      1236  : heat_exhaustion
# 7 HEAT ILLNESS, HEAT STROKE      2466  : heat_stroke
# 8 Heat Stroke                     654  : heat_stroke
# 9 HEAT STROKE                      45  : heat_stroke



# Note: in Paper 2 ("daily HSI"), RMEs listed as "HEAT ILLNESS" (n=128) and "HEAT ILLNESS, HEAT INJURY"(n=1236) were coded as "other" since they are not specified as heat exhaustion or heat stroke.  
# Here, after closer review of "ARMED FORCES REPORTABLE MEDICAL EVENTS | Guidelines and Case Definitions" they will be coded as "heat exhaustion" based on RME inclusion criteria. This is not likely to have impacted the previous findings, as incidence would still be captured under hospitalization or ambulatory records.


df_rme <-
  df %>% 
    dplyr::filter(source %in% "RME") %>% 
    dplyr::mutate(hsi = dplyr::case_when(
         str_detect(dx1_name, regex("stroke", ignore_case = T))
         ~ "heat_stroke",
         TRUE ~ "heat_exhaustion"))
    
df_rme %>% 
    count(hsi)


# Classify Inpatient and Outpatient
  # Inspect classification codes

df %>% 
  dplyr::filter(!source %in% "RME") %>% 
  count(dx1) %>% 
  dplyr::arrange(desc(n))

df %>% 
  dplyr::filter(!source %in% "RME") %>% 
  count(dx2) %>% 
  dplyr::arrange(desc(n))


df_non_rme <-
  df %>% 
   dplyr::filter(!source %in% "RME") %>% 
   dplyr::mutate(hsi = dplyr::case_when(
     str_detect(dx1, "T670XX[A-Z]") | str_detect(dx1, "9920") |
     str_detect(dx2, "T670XX[A-Z]") | str_detect(dx2, "9920")
     ~ "heat_stroke",
     str_detect(dx1, "T67[3459]XX[A-Z]") | str_detect(dx1, "992[345]") |
     str_detect(dx2, "T67[3459]XX[A-Z]") | str_detect(dx2, "992[345]")
     ~ "heat_exhaustion",
        TRUE ~ "other")) 

## Inspect inpatient/outpatient classifications

df_non_rme %>% 
  group_by(source, hsi) %>% 
    count(hsi)




## Combine RME and In/out-patient HSI-classified dataframes

df_hsi <-
  df_non_rme %>% 
    bind_rows(df_rme) %>% 
    mutate(hsi = as_factor(hsi), 
           id = row_number())

```



# Rhabdomyolysis ICD codes
Note: these are rhabdo co-morbidities with other HSI ICD codes and may not represent all rhando cases

```{r}

df %>% 
  dplyr::filter(
      str_detect(dx1, "72888") | str_detect(dx1, "M6282") |
      str_detect(dx2, "72888") | str_detect(dx2, "M6282")) 
  
df_hsi <-
  df_hsi %>% 
    mutate(rhabdo = 
             case_when(
               str_detect(dx1, "72888") | str_detect(dx1, "M6282") |
               str_detect(dx2, "72888") | str_detect(dx2, "M6282") ~ 1,
              TRUE ~ 0)
)


df_hsi %>% 
  count(rhabdo)

```


## Apply case definitions
Limit to one incident case per subject-year

```{r}

# Apply case definition: 
#   considered an incident case only once per calendar year
#   heat stroke is prioritized over heat exhaustion
#   includes only heat stroke and heat exhaustion - will retain "other" in datasets at this stage


# Identify observations to remove: 
#   more than 1 observation per studyid-year
#   heat stroke > heat exhaustion
#   hospitalizations > reportable medical events > ambulatory visits




# Remove `heat_exhaustion` or `other` if `heat_stroke` in studyid/year stratum

clean_non_heat_stroke <-
  df_hsi %>% 
    dplyr::mutate(year = lubridate::year(d_event)) %>% 
    janitor::get_dupes(studyid, year) %>% # extract sets of instances with > 1 obs per studyid per year
    group_by(studyid, year) %>% 
    dplyr::filter(any(hsi == "heat_stroke")) %>% # filter strata that include a heat stroke obs
    dplyr::filter(!hsi %in% "heat_stroke") %>% # filter all non- heat stroke obs
    dplyr::ungroup() %>% 
    dplyr::select(-dupe_count, -year)

clean_non_heat_stroke # obs to be removed


clean_df_step1 <-
    df_hsi %>% 
      dplyr::setdiff(clean_non_heat_stroke)  # remove rows that appear in `clean_non_heat_stroke`

# Inspect

clean_df_step1 %>% 
  count(hsi)


 # Prioritize Hospitalizations
 
clean_prioritize_hosp <-
  clean_df_step1 %>% 
      dplyr::mutate(year = lubridate::year(d_event)) %>% 
        janitor::get_dupes(studyid, year) %>% # extract sets of instances with > 1 obs per studyid per year
        group_by(studyid, year) %>% 
        dplyr::filter(any(source == "INPATIENT")) %>% 
        dplyr::filter(!source %in% "INPATIENT") %>%
        dplyr::ungroup() %>% 
        dplyr::select(-dupe_count, -year)

clean_prioritize_hosp # obs to be removed

clean_df_step2 <-
    clean_df_step1 %>% 
      dplyr::setdiff(clean_prioritize_hosp) 

# Prioritize RME over ambulatory
clean_prioritize_rme <-
  clean_df_step2 %>% 
      dplyr::mutate(year = lubridate::year(d_event)) %>% 
        janitor::get_dupes(studyid, year) %>% 
        group_by(studyid, year) %>% 
        dplyr::filter(!any(source == "INPATIENT")) %>% # exclude strata with inpatient
        dplyr::filter(any(source == "RME")) %>%
        dplyr::filter(!source %in% "RME") %>%
        dplyr::ungroup() %>% 
        dplyr::select(-dupe_count, -year)

clean_df_step3 <-
    clean_df_step2 %>% 
      dplyr::setdiff(clean_prioritize_rme) 


# Retain first observation in subjectid-year stratum

drop_repeat_vists <-
  clean_df_step3 %>% 
    dplyr::mutate(year = lubridate::year(d_event)) %>% 
    janitor::get_dupes(studyid, year) %>%
    dplyr::group_by(studyid, year) %>%
    dplyr::arrange(d_event) %>% 
    dplyr::filter(!row_number() == 1) %>% 
    dplyr::ungroup() %>% 
    dplyr::select(-dupe_count, -year)

# Join with `persons` df for bmi
casedef <-  
  clean_df_step3 %>% 
    dplyr::setdiff(drop_repeat_vists) %>% 
    dplyr::left_join(persons, by = "studyid")


# Match with full dataset with new column for casedef
# Add BMI categories
# Add year and decade

df_hsi <-
  df_hsi %>% 
    mutate(casedef = if_else(id %in% casedef$id, 1, 0),
           bmi_cat =
               as.factor(case_when(
                 bmi < 18.5 ~ "underweight",
                 bmi >= 18.5 & bmi < 25 ~ "normal",
                 bmi >= 25 & bmi < 30 ~ "overweight",
                 bmi >= 30 ~ "obese"
               )),
           bmi_cat = fct_relevel(bmi_cat, 
                "underweight", "normal", "overweight", "obese"),
           age = fct_relevel(age, ">=40", after = Inf),
           rhabdo = as.factor(rhabdo),
           casedef = as.factor(casedef),
           year = lubridate::year(d_event),
           decade = as.factor(floor(year / 10) * 10)) %>% 
   mutate_if(is.character, as.factor) 
  
           
           
   
df_hsi %>% count(casedef)

#write_rds(df_hsi, "data/df_hsi.rds")




```



```{r}

# df_hsi <- read_rds("data/df_hsi.rds")

Hmisc::describe(df_hsi)

```


## Table 1


### Labels
```{r}

# Labels for Table

label(df_hsi$hsi) <- "HSI Diagnosis"
label(df_hsi$age) <- "Age Group"
label(df_hsi$sex) <- "Sex"
label(df_hsi$race_ethnic) <- "Race/Ethnicity"
label(df_hsi$grade) <- "Grade Group"
label(df_hsi$service) <- "Service Branch"
label(df_hsi$bmi) <- "BMI"
label(df_hsi$rhabdo) <- "Rhabdomyolysis Co-morbidity"
label(df_hsi$casedef) <- "Case-defined HSI (unique incidence)"
label(df_hsi$bmi_cat) <- "BMI Categories"
label(df_hsi$decade) <- "Decade of Incidence"






```




All
```{r}

## Outcome table
```{r}



to_table1 <-
  df_hsi %>% 
    dplyr::select(source, age, sex, race_ethnic, service, grade, hsi, bmi, rhabdo, casedef, bmi_cat, decade, year) %>% 
       mutate(service = recode(service,
            "A" = "Army",
            "C" = "Coast Guard",
            "F" = "Air Force",
            "M" = "Marine Corps",
            "N" = "Navy"),
            source = recode(source,
            "INPATIENT" = "In-Patient",
            "OUTPATIENT" = "Out-Patient",
            "RME" = "Reportable Event"),
            hsi = recode(hsi,
            "heat_exhaustion" = "Heat Exhaustion",
            "heat_stroke" = "Heat Stroke"),
            sex = recode(sex,
            "F" = "Female",
            "M" = "Male",
            "Z" = "Unknown")
            ) 



table1::table1(~ hsi + sex + age + sex + grade + race_ethnic + rhabdo + bmi + bmi_cat + casedef + decade | source, data = to_table1)



```


## Casedef only

```{r}

to_table1_casedef <-
  df_hsi %>% 
    filter(casedef == 1) %>% 
    dplyr::select(source, age, sex, race_ethnic, service, grade, hsi, bmi, rhabdo, bmi_cat, decade, year) %>% 
       mutate(service = recode(service,
            "A" = "Army",
            "C" = "Coast Guard",
            "F" = "Air Force",
            "M" = "Marine Corps",
            "N" = "Navy"),
            source = recode(source,
            "INPATIENT" = "In-Patient",
            "OUTPATIENT" = "Out-Patient",
            "RME" = "Reportable Event"),
            hsi = recode(hsi,
            "heat_exhaustion" = "Heat Exhaustion",
            "heat_stroke" = "Heat Stroke"),
            sex = recode(sex,
            "F" = "Female",
            "M" = "Male",
            "Z" = "Unknown")
            ) 


table1::table1(~ hsi + sex + age + sex + grade + race_ethnic + rhabdo + bmi + bmi_cat + decade | source, data = to_table1_casedef)

table1::table1(~ hsi + sex + age + sex + grade + race_ethnic, data = to_table1_casedef)

```






## BMI Tables

```{r}

# All

table1::table1(~ hsi + sex + age + sex + grade + race_ethnic + rhabdo + casedef + decade | bmi_cat, data = to_table1)


table1::table1(~bmi_cat | decade, data = to_table1)

```


## BMI barplots, all observations
```{r}
# All observations, by decade

# dodged bars

to_table1 %>% 
  ggplot(aes(x = decade, fill = bmi_cat)) +
    geom_bar(position = position_dodge()) +
    theme_bw()

# relative proportions at each decade/bmi by stacking the bars and then standardising each bar to have the same height

to_table1 %>% 
  ggplot(aes(x = decade, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()

# Remove NA's

to_table1 %>%
  filter(!bmi_cat %in% NA) %>% 
  ggplot(aes(x = decade, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()



# All observations, by year

to_table1 %>% 
  ggplot(aes(x = year, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()

to_table1 %>%
  filter(!bmi_cat %in% NA) %>% 
  ggplot(aes(x = year, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()


```



## BMI barplots, case defined HSI

```{r}

# dodged bars

to_table1_casedef %>% 
  ggplot(aes(x = decade, fill = bmi_cat)) +
    geom_bar(position = position_dodge()) +
    theme_bw()

# relative proportions at each decade/bmi by stacking the bars and then standardising each bar to have the same height

to_table1_casedef %>% 
  ggplot(aes(x = decade, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()

# Remove NA's

to_table1_casedef %>%
  filter(!bmi_cat %in% NA) %>% 
  ggplot(aes(x = decade, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()



# All observations, by year

to_table1_casedef %>% 
  ggplot(aes(x = year, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()

to_table1_casedef %>%
  filter(!bmi_cat %in% NA) %>% 
  ggplot(aes(x = year, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()

```


## BMI: Restrict to Age < 25

```{r}

# dodged bars

to_table1_casedef %>%
  filter(age %in% c("<20", "20-24"),
          year < 2020) %>% 
  ggplot(aes(x = decade, fill = bmi_cat)) +
    geom_bar(position = position_dodge()) +
    theme_bw()

# relative proportions at each decade/bmi by stacking the bars and then standardizing each bar to have the same height

to_table1_casedef %>% 
    filter(age %in% c("<20", "20-24")) %>% 
  ggplot(aes(x = decade, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()

# Remove NA's

to_table1_casedef %>%
  filter(age %in% c("<20", "20-24"),
         !bmi_cat %in% NA,
         year < 2020) %>% 
  ggplot(aes(x = decade, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()



# All observations, by year

to_table1_casedef %>%
    filter(age %in% c("<20", "20-24"),
         year < 2020) %>% 
  ggplot(aes(x = year, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()


to_table1_casedef %>%
    filter(age %in% c("<20", "20-24"),
         !bmi_cat %in% NA,
         year < 2020) %>% 
  ggplot(aes(x = year, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()

# boxplot

to_table1_casedef %>%
    filter(age %in% c("<20", "20-24"),
         !bmi_cat %in% NA,
         year < 2020) %>% 
  ggplot(aes(x = as.factor(year), y = bmi)) +
    geom_boxplot() +
    theme_bw()



# scatterplot

to_table1_casedef %>%
  filter(age %in% c("<20", "20-24"),
         !bmi_cat %in% NA,
         year < 2020) %>%
ggplot(aes(x = year, y = bmi)) +
    geom_point() +
    geom_smooth(method = "lm") +
    theme_bw()


# Linear model

to_table1_casedef %>%
    filter(age %in% c("<20", "20-24"),
         !bmi_cat %in% NA,
         year < 2020) %>%
  lm(bmi ~ year, data = .) %>% 
  broom::tidy()

```



## BMI: Restrict to Age < 20

```{r}

# dodged bars

to_table1_casedef %>%
  filter(age %in% "<20") %>% 
  ggplot(aes(x = decade, fill = bmi_cat)) +
    geom_bar(position = position_dodge()) +
    theme_bw()

# relative proportions at each decade/bmi by stacking the bars and then standardizing each bar to have the same height

to_table1_casedef %>% 
    filter(age %in% "<20") %>% 
  ggplot(aes(x = decade, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()

# Remove NA's

to_table1_casedef %>%
  filter(age %in% "<20",
         !bmi_cat %in% NA,
         year < 2020) %>% 
  ggplot(aes(x = decade, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()



# All observations, by year

to_table1_casedef %>%
    filter(age %in% "<20",
         year < 2020) %>% 
  ggplot(aes(x = year, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()


to_table1_casedef %>%
    filter(age %in% "<20",
         !bmi_cat %in% NA,
         year < 2020) %>% 
  ggplot(aes(x = year, fill = bmi_cat)) +
    geom_bar(position = position_fill()) +
    theme_bw()



to_table1_casedef %>%
    filter(age %in% "<20",
         !bmi_cat %in% NA,
         year < 2020) %>% 
  ggplot(aes(x = as.factor(year), y = bmi)) +
    geom_boxplot() +
    theme_bw()

```
