==> 2_demographics.Rmd <==
---
title: "demographics"
author: "SL"
date: "5/23/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(haven)
library(Hmisc)
library(janitor)
library(skimr)
library(zoo)
library(naniar)
library(readxl)
#library(simstudy)


```



Import demographics from DMED queries

```{r}

setwd("C:/Users/slewa/OneDrive - cumc.columbia.edu/risk_factors")

demog <-
  readxl::read_xlsx(path = "./data/dmed_demog.xlsx", sheet = "all",
                    col_names = TRUE) %>% 
  mutate(year = as.factor(year))

```


Wrangle from wide to long



```{r}

# proportion for each variable-category/year
# population sum for each category-year



demog_long <-
  demog %>% 
    pivot_longer(-year, names_to = "var", values_to = "population") %>% 
      mutate(var = as.factor(var),
             cat = as.factor(str_replace_all(var, "[:digit:]", ""))) %>% 
  nest(data = c(var, population)) %>% 
  mutate(data = 
           map(.x = .$data, ~ .x %>% 
                      mutate(proportion = population / sum(population),
                             sum_pop = sum(population)))) %>% 
  unnest(cols = c(data))
  


# Note: population sums vary slightly between categories in a given year
  # race and service match with the highest counts (use service population total)

demog_long %>% 
  group_by(year, cat) %>% 
  summarise(sum_pop = mean(sum_pop)) 

# write_rds(demog, file = "data/demog.rds")
# write_rds(demog_long, file = "data/demog_long.rds")


```




## Simulate active duty population for each year

Note: limitation of simulated population --
  We do not know the correlations between variables (will be strong between age and rank; these variables should be evaluated independently)

```{r}


demographics <- 
  demog_long %>%
    nest(data = -year) %>% 
    mutate(pop = map(data, ~ (.) %>% filter(var %in% "service1") %>% pull(sum_pop) %>% round(., digits = 0))) %>% 
  unnest(pop)

# Note: total population based on "service" used for all categories each year



set.seed(82)

sim_fun = function(df, pop) {
    data.frame(
      age = sample(x= 1:6, size = pop, replace = TRUE, prob = c(
        df %>% filter(var %in% "age1") %>% pull(proportion),
        df %>% filter(var %in% "age2") %>% pull(proportion),
        df %>% filter(var %in% "age3") %>% pull(proportion),
        df %>% filter(var %in% "age4") %>% pull(proportion),
        df %>% filter(var %in% "age5") %>% pull(proportion),
        df %>% filter(var %in% "age6") %>% pull(proportion)
        )),
      sex = sample(x = 1:2, size = pop, replace = TRUE, prob = c(
        df %>% filter(var %in% "sex1") %>% pull(proportion),
        df %>% filter(var %in% "sex2") %>% pull(proportion)
         )),
      grade = sample(x = 1:4, size = pop, replace = TRUE, prob = c(
        df %>% filter(var %in% "grade1") %>% pull(proportion),
        df %>% filter(var %in% "grade2") %>% pull(proportion),
        df %>% filter(var %in% "grade3") %>% pull(proportion),
        df %>% filter(var %in% "grade4") %>% pull(proportion)
      )),  
      race = sample(x = 1:3, size = pop, replace = TRUE,prob = c(
        df %>% filter(var %in% "race1") %>% pull(proportion),
        df %>% filter(var %in% "race2") %>% pull(proportion),
        df %>% filter(var %in% "race3") %>% pull(proportion)
      )),  
      service = sample(x = 1:4, size = pop, replace = TRUE, prob = c(
        df %>% filter(var %in% "service1") %>% pull(proportion),
        df %>% filter(var %in% "service2") %>% pull(proportion),
        df %>% filter(var %in% "service3") %>% pull(proportion),
        df %>% filter(var %in% "service4") %>% pull(proportion)
    ))
  )
}


sim_df <-
  demographics %>% 
    mutate(sim_pop = 
      map2(.x = demographics$data, .y = demographics$pop, .f = sim_fun))



sim_df


# write_rds(sim_df, file = "data/sim_df.rds")

# sim_df <- read_rds(file = "data/sim_df.rds")
```


## Inspect sim_pop
```{r}

sim_df$sim_pop[[1]] %>% 
  head()


sim_df$sim_pop[1] %>% 
  as.data.frame() %>% 
  as_tibble()

# Sim age proportions

sim_df$sim_pop[1] %>% 
  as.data.frame() %>% 
  as_tibble() %>% 
  mutate_all(factor) %>% 
  count(age) %>% 
  mutate(sim_prop = n / sum(n))

# Compare to base demographics

demog_long %>% 
  filter(year %in% "1990" & cat %in% "age")
```
