==> 3_attributal_risk.Rmd <==
---
title: "attributal_risk"
author: "SL"
date: "6/01/2021"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(haven)
library(Hmisc)
library(janitor)
library(skimr)
library(zoo)
library(naniar)
library(broom)
library(MASS)
library(viridis)

memory.limit(size = 50000)
# rm(list = ls())
```


```{r}

df_hsi <- read_rds("data/df_hsi.rds")
demog <- read_rds("data/demog.rds")
demog_long <- read_rds("data/demog_long.rds")
sim_df <- read_rds(file = "data/sim_df.rds")
```



```{r}

hsi_coded <-
  df_hsi %>% 
    dplyr::select(year, age, sex, race_ethnic, service, grade, casedef, hsi) %>% 
    mutate(
      age = recode(
        age, 
          "<20" = "age1",
          "20-24" = "age2",
          "25-29" = "age3",
          "30-34" = "age4",
          "35-39" = "age5",
          ">=40" = "age6"),
      sex = recode(
        sex,
          "M" = "sex1",
          "F" = "sex2",
          "Z" = "sex_NA"),
      race_ethnic = recode(
         race_ethnic,
          "American Indian/Alaskan Native" = "race3",
          "Asian/Pacific Islander" = "race3",
          "Hispanic" = "race3",
          "Non-Hispanic Black" = "race2",
          "Non-Hispanic White" = "race1",
          "Other" = "race3",
          "Unknown" = "race_NA"), 
      grade = recode(
         grade,
          "E1-E4" = "grade1",
          "E5-E9" = "grade2",
          "O1-O3/W1-W3" = "grade3",
          "O4-O10/W4-W5" = "grade4"), 
      service = recode(
         service,
          "A" = "service1",
          "N" = "service2",
          "F" = "service3",
          "M" = "service4",
          "C" = "service_CG")  #"C" coded as NA
    ) %>% 
  rename(race = race_ethnic)


hsi_coded


```
  

```{r}
demog
demog_long

# Filter to remove Coast Guard, and select only case-defined encounters (heat stroke, heat exhaustion)

# pivot long to "cat" and "var"
# count by year


hsi_long <-
  hsi_coded %>%
    filter(!service %in% "serviceCG",
           !hsi %in% "other", # 
           casedef == 1) %>%
    dplyr::select(-casedef, -hsi) %>%
    mutate(across(everything(), as.character)) %>% 
    pivot_longer(-year, names_to = "cat", values_to = "var") %>% 
    group_by(year, var) %>% 
    summarise(n = n()) %>% 
    mutate(cat = str_replace_all(var, "[:digit:]", ""),
           cat = str_replace(cat, "_.*", "")) %>% 
  filter(!cat %in% NA)


hsi_long

hsi_long %>% ungroup() %>% count(cat)


# Calculate proportion for each variable/year

hsi_prop <-
  hsi_long %>% 
    nest(data = c(var, n)) %>%
    mutate(data = 
             map(data, ~ .x %>% 
                        mutate(prop_hsi = n / sum(n),
                               sum_hsi = sum(n)))) %>% 
    unnest(cols = c(data)) 

hsi_prop
```


## Combine HSI and demog dataframes

```{r}

hsi_rf <-
  demog_long %>% 
    dplyr::select(-cat) %>%  # joing table has cat column
    left_join(hsi_prop, by = c("year", "var")) %>% 
  mutate(risk = n / population,
         prop_ratio = prop_hsi / proportion)

hsi_rf

```


## Risk over time

```{r}


hsi_rf %>% 
  mutate(var_num = str_extract(var, "[:digit:]")) %>% 
  ggplot(aes(x = as.integer(year), y = risk, color = cat, shape = var_num)) +
    geom_point() +
    geom_line() +
  theme_bw()


# heatmap 

hsi_rf %>% 
  ggplot(aes(x = as.integer(year), y = var, fill = risk)) +
  geom_tile() +
  scale_fill_viridis(discrete = FALSE) +
  theme_bw()

```



## Risk factor proportion

```{r}

hsi_rf %>% 
  mutate(var_num = str_extract(var, "[:digit:]")) %>% 
  ggplot(aes(x = as.integer(year), y = prop_ratio, color = cat, shape = var_num)) +
    geom_point() +
    geom_line() +
  theme_bw()


# heatmap

hsi_rf %>% 
  ggplot(aes(x = as.integer(year), y = var, fill = prop_ratio)) +
  geom_tile() +
  scale_fill_viridis(discrete = FALSE) +
  theme_bw()


```




## Statistical Tests


Chi-square goodness of fit: Compare multiple observed proportions to expected probabilities


Null hypothesis (H0): There is no significant difference between the observed and the expected value.
Alternative hypothesis (Ha): There is a significant difference between the observed and the expected value.


simulate.p.value = TRUE, a logical indicating whether to compute p-values by Monte Carlo simulation.
fisher.test(a)

```{r}

chisq <-
  hsi_rf %>% 
    nest(data = c(-year, -cat)) %>% 
    mutate(chisq =
             map(data,
             ~chisq.test(x = .$n, p = .x$proportion))
  )




chisq_sim <-
  hsi_rf %>% 
    nest(data = c(-year, -cat)) %>% 
    mutate(chisq =
             map(data,
             ~chisq.test(x = .$n, p = .x$proportion,
                         simulate.p.value = TRUE, B = 2000))
  )


# write_rds(chisq_sim, file = "data/chisq_sim")

chisq_sim <- read_rds(file = "data/chisq_sim")


chisq_sim <-
  chisq_sim %>% 
    mutate(chisq_tidy = map(chisq, broom::tidy))


chisq_sim


# Chisq significance stars

chisq_sim %>% 
  unnest(chisq_tidy) %>%
  dplyr::select(year, cat, statistic, p.value) %>% 
   mutate(signif = case_when(
        p.value <= 0.05 & p.value > 0.01 ~ "*",
        p.value <= 0.01 & p.value > 0.001 ~ "**",
        p.value <= 0.001 ~ "***",
        TRUE ~ ""
    )) 

# Chisq significance, wide table by year
chisq_sig_wide <-
  chisq_sim %>% 
    unnest(chisq_tidy) %>%
    dplyr::select(year, cat, statistic, p.value) %>% 
    mutate(signif = case_when(
        p.value <= 0.05 & p.value > 0.01 ~ "*",
        p.value <= 0.01 & p.value > 0.001 ~ "**",
        p.value <= 0.001 ~ "***",
        TRUE ~ ""
    )) %>%
    dplyr::select(year, cat, signif) %>% 
    pivot_wider(names_from = year, values_from = signif) 


chisq_sig_wide 

chisq_sig_wide %>% 
  View()
         




```
